criterion float param

+++ src/losses/criterion.py
@@ -6,4 +6,4 @@ TverskyLoss = smp.losses.TverskyLoss(mode='multilabel', log_loss=False)
 BCELoss     = smp.losses.SoftBCEWithLogitsLoss()
 
 def criterion(pr, gt):
-    return 0.5*BCELoss(pr, gt) + 0.5*TverskyLoss(pr, gt)
+    return 0.5*BCELoss(pr, gt.float()) + 0.5*TverskyLoss(pr, gt)
